Add files via upload
PROBLEM STATEMENT : In the byteland country a string "s " is said to be super ascii if and only if the count of each character of the string is equal to its ascii value. In the byteland country, the ascii code of 'a' is 1, 'b' is 2........and 'z' is 26. Your task is to find out the string is super ascii or not.
INPUT FORMAT : 
The first line contains the number of testcases T , followed by T lines, each line contains a string 's'.
OUTPUT FORMAT :
For each test case, print YES if the string is super ascii else print NO